Replaced all calls to '.flat' with use of 'ravel()' method to support display of subarrays or sections.
Changed interface to 'set()' by renaming 'scale' to 'transform', changing 'factor' to 'offset', and adding 'scale' as a multiplicative factor as well. The 'transform' parameter now accepts a Python function directly and applies it to the array with a no-op default transformation built-in.

'_scaleImage' renamed '_transformImage' and rewritten to use 'self.transform directly on array with offset into scale. '_transformImage' renamed '_fbclipImage' to reflect its actual operations.

The module function 'bscaleImage' was moved into the class as a method. The pmin and pmax values changed to 'float's to support proper computation of scale value for INT images.

The 'display' method also now accepts all 'set()' parameters as well and passes them to the 'set()' method.

Bounds checking done on transformation so that if there was an error during the transformation, it will simply resort to using the original input array along with full pixel value range.

Documentation was updated to reflect changes.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@13 fe389314-cf27-0410-b35b-8c050e845b92