Bugfix: corrected handling of case when calcimage != targimage.
Previously in this case, the output FITS file consisted of the calcimage file with only the data from the first extension of the targimage SAA-cleaned and inserted in the right place.

This version produces an outimage that is correctly identical with the targimage except for the SAA-cleaned data and SAACLEAN-inserted keywords in the primary header.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@502 fe389314-cf27-0410-b35b-8c050e845b92