Bugfix: corrected logic when calcimage != targimage, so the data is handled correctly.
(Previously, the targimage data was only preserved for pixels when no correction was needed.)

In making this fix, also laid some groundwork to make it easier for the task to produce a diagnostic file consisting of exactly the persistence signal that was removed (flat-fielded, scaled, only the affected pixels). It's not written out yet, but it is now calculated.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@503 fe389314-cf27-0410-b35b-8c050e845b92