wavecal.py and concurrent.py were modified to add together the NUV stripes (and to add together the template spectra) before taking the cross correlation to find the offset.  There are loops over segments (stripes for NUV), and previously the cross correlation was done within the loops.  The code to print the results has also been modified.  See OPR 57228.
The TAGFLASH primary header keyword was originally boolean, T or F.  The
allowed values are now strings:  "NONE", "AUTO", "UNIFORMLY SPACED".
"UNIFORMLY SPACED" is an engineering mode that can result in many flashes,
in particular more flashes than there are LMP_ONi (and other) keywords,
and concurrent.py has been modified to process all of them (the total
number is given by NUMFLASH) but only update the values of keywords for
index i if LMP_ONi already exists in the header.  See OPR 56407 for
further details.

Many of the boolean variables that were previously assigned values
of 1 or 0 are now assigned True or False.

In calcos.py, the code for updating the boolean MEMPRSNT column in the
association table was modified.  The numpy version of pyfits no longer
uses data._parent (although the numarray version still does), so the
value in the MEMPRSNT column for the product is now set directly to True
after writing all the output files.

The return value for getTable in cosutil.py will now be None if there are
no rows matching the filter (unless exactly_one=True or at_least_one=True,
in which case getTable will still raise an exception).  There are some
modes with no spectral data (e.g. for NUV stripe C), and it seemed more
reasonable for getTable to return None when there is no matching row in
a reference table.

When processing a wavecal, aperture is now explicitly set to "WCA" for
selecting rows from a reference table.  The value of the APERTURE header
keyword might be set to "PSA" (especially for thermal vac data), and in
that case the results would be wrong, due to the offset between those two
apertures.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@592 fe389314-cf27-0410-b35b-8c050e845b92