Changes to accum.py:
In globrate_accum, the activeArea function was misspelled.
In findStimAccum, the setshape method was being called, but that does
not exist in numpy.

Changes to calcos.py:

There was a calling-sequence error in the call to calcos; one of the
optional arguments was missing.  This was fixed by using keyword=value
syntax.

In sanityCheck, there is a check for an inconsistency between OBSTYPE
and OPT_ELEM.  For imaging observations the optical element can be
MIRRORA or MIRRORB, but during thermal vac the corresponding values
were TA1Image and TA1Brght respectively.  Calcos was only checking
for the latter values.
The following:
        if info["obstype"] == "SPECTROSCOPIC" and \
           info["opt_elem"][0:3] == "TA1":
has been changed to:
        if info["obstype"] == "SPECTROSCOPIC" and \
               (info["opt_elem"][0:6] == "MIRROR" or
                info["opt_elem"][0:3] == "TA1"):


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@1017 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: e5efdbd0884f100abc2dc8f8b6ee50c5ad479011