PR 58936 Trac ticket #133
Changes to calcos.py:

The RPTCORR switch is now ignored.  Averaging the x1d tables together
doesn't affect any other output product, and the output tables are small,
so there didn't seem to be any point in having to check a switch before
averaging the data.  Also, the keyword rules currently set RPTCORR to OMIT.

Variable names referring to the FPPOS index (header keyword, 1..4) have
been renamed to replace "osm" with "fppos".  This change was made just
for clarity.

Change to accum.py, burst.py, concurrent.py, cosutil.py, timetag.py:

The activeArea function in cosutil.py was modified to return all four
borders (bottom, top, left, right) of the active area.  Currently the
left and right borders are not used.  The other files were modified
because they call activeArea.

Change to accum.py:

The pulse-height parameters are now multiplied by four before using them
to check that the pulse-height histogram for accum data is reasonable.
The reason for this change is that the pulse heights are in the range
0 to 31, but the pulse-height histograms are scaled to 0 to 127.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@1062 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: f8a3a52cb56545870ec699669e97cad1b31ffdef