Added the following functionality:
- fileutil:
when a GEIS file is passed to openImage it will convert to MEF format
not only the science file but also the data quality file if it's
present.

- wcsutil:
Added a new keyword to the write method - wcs=None. It allows to pass
a WCSObject with WCS values to be used for updating the
image. This is meant to be used with GEIS files 
to allow to update the WCS keywords in the header of the
data quality file with the values of the science file. 
Resolves #157,

- makewcs:
Resolves #157 - updates the WCS keywords in the header of a WFPC2 data
quality file to be consistent with the WCS keywords in the science
file.



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@1158 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 93faa3e928c5cb669eb58eca6d2eb5bb6b9a8a63