Changes to accum.py:
In accum.doRandcorr, addRandomNumbers now returns a value, the seed that was
actually used to initialize the pseudorandom number generator.  The primary
header keyword RANDSEED for the output file will be updated with the value
returned by addRandomNumbers.

Delete lines 740-742:
    region.ravel()
    x_region.ravel()
    y_region.ravel()

Changes to calcos.py:

Call getinfo.resetSwitches in getHeaderInfo.

A given wavecal exposure can be included in more than one association,
and the name of the wavecal is the same in each association.  So after
running calcos on one such association, calcos would not calibrate another
because some output files are already present on disk.  Modified to print
a warning and delete those output wavecal files.

Change to calcosparam.py:

Added data quality flags DQ_DIVOTS = 8192 and DQ_SDISTORTION = 16384.

Change to cosutil.py:

Modified addRandomNumbers to use the numpy.random function to generate
pseudo-random numbers (instead of ccos.addrandom).  addRandomNumbers
now returns the value that was actually used as a seed (only different
from the input seed if the input seed was -1).

Change to extract.py:

Function postargOffset was added; this computes the offset of the spectrum
in the cross-dispersion direction resulting from a postarg.  doExtract was
modified to call this, adding the returned value to shift2.

Change to fpavg.py:

In computeOutputInfo, the output array for NUV data is now made one element
larger to avoid indexing problems due to roundoff.  (FUV uses nearest
neighbor; there is no interpolation.)  The problem caused by roundoff is
that when interpolating in accumulateSums, if there is really no offset
between input and output arrays, it can be that weights p and q will be
1 and some tiny number (e.g. 3.e-13) respectively.  Since the offset was
zero, the output array wasn't large enough to accomodate the input array
shifted by one pixel (and multiplied by q, i.e. by essentially zero).

Change to getinfo.py:

Added function resetSwitches, which sets a calibration switch to "SKIPPED"
if any reference file needed by the step has name "N/A".  Currently,
BADTCORR is the only switch in the list of switches that can be reset.

Changes to timetag.py:

In computeThermalParam, if a stim is not found within some time interval,
the stim position from the previous time interval will be used.  For the
first time interval, the default value is the reference position.

In writeImages when binning events to the output image, events will be
excluded depending on the data quality flags.  There are many regions on
the detector with non-zero data quality flags, however, and it really is
not appropriate to exclude events in all of those regions.  The DQ
extension retains the information about which regions are flagged, and
that should be sufficient.  So now events are excluded only if the data
quality flag includes at least one of the following:  burst, within a
bad-time interval, or the pulse height is out of bounds.

In doRandcorr, addRandomNumbers now returns a value, the seed that was
actually used to initialize the pseudorandom number generator.  The primary
header keyword RANDSEED for the output file will now be updated with the
value returned by addRandomNumbers.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@1197 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 4524a51e8a760a9531d2c27c0e0279df4c463708