Numdisplay Version 1.5 (11-July-2008): This version addresses a number of issues; namely, no longer directly dependent on PyRAF (Ticket #199), new modules added from Gemini's PyGACQ package, socket management was serialized (Ticket #224), automatic reset of connection to ds9 (Ticket #222).
The package structure of 'numdisplay' was also revised to put all the source code in a 'lib' subdirectory to be consistent with the rest of the 'stsci_python' packages (Ticket #223).  

The new modules which have been added provide for overlay capabilities similar to 'tvmark', although a bit more limited.  The overlay module was updated to include a new method for drawing markers (characters) based on the 'ichar' module and bit-mapped font derived from IRAF's 5x7 stroke font.    




git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@6801 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 5694bc9954378b8a591408719e3976be802db3da