In missingRefFiles in calcos.py, check both time-tag and accum data, set a switch to perform if it's perform for either type of data, and override a reference file name if it's "N/A" for one type but not the other. This still doesn't handle the possibility that the user might have specified different reference files in different raw files, but that shouldn't happen in the pipeline.
In copySptFile in calcos.py, make sure the spt file is for a science exposure
rather than a wavecal (unless there is no spt for any science exposure).
Change the suffix from "spt" to "jnk", i.e. the name of the copy will be
<product>_jnk.fits.  Remove the code that updated extension keywords.
Add two comment records to the primary header to explain that the file
can be ignored by the user.

In processWavecal in calcos.py, there is a call to printWavecalRef to
print the names of the reference files used for wavecal processing.  This
function was called with a list of reference files which might be for a
science observation rather than a wavecal.  Even though the file names
ought to be the same, the call to printWavecalRef was moved into the loop
to ensure that the reference file names that are printed are actually
for a wavecal.

In findShifts in concurrent.py, the following is incorrect:
        if cosutil.findColumn (disptab, "fpoffset"):
            filter_lamp["fpoffset"] = self.info["fpoffset"]
The test should be on lamptab, not disptab.

See PRs 60653, 60708, Trac tickets 243, 247.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@7009 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 7b712adc87814a3124d6aa1cf7797841645ef424