Changes to cosutil.py:
The test for imaging data in getInputDQ checked the value of obsmode,
and it should have checked obstype.  This resulted in an incorrect
size for the dq array for NUV IMAGING data.  Trac #345, PR 61955

Changes to extract.py:

In updateArchiveSearch, check for the existence of the DQ_WGT column
by calling cosutil.findColumn, rather than using a try/except block.
The problem with the latter is that pyfits used to raise a NameError
exception for a missing column, but now it raises a KeyError, and not
everyone has a recent enough version of pyfits.  Trac #346, PR 61957

Changes to getinfo.py:

The size of calibrated NUV IMAGING data should be only 1024x1024,
not 1274x1024.  The test for imaging data in getGeneralInfo checked
the value of obsmode, however, and it should have checked obstype.
Trac #345, PR 61955

Changes to timetag.py:

In timetagBasicCalibration, the check on whether there are any rows in
the table was to get the length of ofd["EVENTS"].data, but if there are
no rows, that will be None.  This was changed to explicitly check whether
ofd["EVENTS"].data is None.  Trac #344, PR 61954

In getWavecalOffsets, check whether active_area actually has any
True elements; if not, return (0, 0, 0, 0).  Trac #343, PR 61953


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@7573 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 0d9436be8ed9cc7ce52a62647bdeb344400a631f