Change to cosutil.py:
In flagOutOfBounds, delete the statements to return if the subarray flag
(keyword SUBARRAY) is false or if the number of subarrays (keyword NSUBARRY)
is zero.  Modify the "if not subarrays:" section to make it more like the
"for i in indices:" section; the problem was that x1 and y1 should have had
xwidth and ywidth subtracted.  Add x_offset to the calling sequence for
flagOutsideActiveArea, and add it to b_left and b_right; this function is
only relevant for FUV data, but if x_offset for FUV is later changed to a
non-zero value, this will reduce the chance of a bug.  Trac #385, PR 62233

Change to fpavg.py:

Delete 'assert nfiles >= 1'.  In oneInputFile, call cosutil.modifyAsnMtyp.
Trac #386

Change to timetag.py:

In doDqicorr, call cosutil.flagOutsideActiveArea only for FUV data, and
add x_offset to the calling sequence.  Trac #385, PR 62233


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@7791 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 0cbc8a79d172e799a2189a12edf95b63d27bfb3a