Changes to calcos.py:
If there are no rawtag or rawaccum files with the specified rootname,
and if there are no rawacq that are ACQ/IMAGE, the warning message
used to be:
There are no raw files for `rootname'
and this has been changed to:
There are no files to calibrate for 'rootname'
If this occurs for the last association or the last file specified,
calcos.py will exit with return code 5 (instead of 0).
Trac #396, PR 62435

In processWavecal, the call to wavecal.storeWavecalInfo now includes
the raw file name.
Trac #395, PR 62434

In function mergeKeywords, keywords CHI_SQ_A, CHI_SQ_B, NDF_A, NDF_B are
included in the list of keywords to be copied from segment A headers to
segment B headers or vice versa.  Trac #394, PR 62433

Changes to concurrent.py:

Function findShifts was modified to use findshift1.  If the fp_pixel_shift
column is found in the lamptab, that value will be added to the shift in
the dispersion direction for each stripe.
If the spectrum is not found in the cross-dispersion direction, more
information is now printed to the trailer, including "----" as the shift,
and the note "# not found in XD".
Function setShiftKeywords now finds the sum of the Chi square values and
the sum of the number of degrees of freedom for each segment or stripe,
and these are assigned to keywords.
Function outFlashSetup also creates columns CHI_SQUARE and N_DEG_FREEDOM.
Trac #393, 394, PR 62432, 62433

In findShifts, removed the code that checks for the existence of the xd_range
and box columns in the wcptab and assigns default values if they are not
found.  That is, these columns are now required, since they are in the
wcptab tables in CDBS.  Trac #332, PR 61783

Change to cosutil.py:

In the section for tempcorr in flagOutOfBounds, for TIME-TAG data there
will typically be more than one set of linear distortion parameters, so
instead of just taking the first set, all the parameters will be averaged,
and the average value will be used for the distortion of the DQ array.
Trac #397, PR 62436

In printMode, also print the value of fpoffset for spectroscopic data.
Trac #398, PR 62437

A new function precess was added.  This is currently not used.  It was
copied from calstis, where it is called in the code that computes the
heliocentric radial velocity, but the difference due to precession is
negligible.
Trac #400, PR 62439

Changes to extract.py:

In copyKeywordsToInput, the list of keywords has been broken into two
sets, one for FUV and one for NUV, so as to avoid putting _c keywords
into FUV headers.  Trac #403, PR 62480

In concatenateFUVSegments, if both segments are present and have rows,
the primary header keyword SEGMENT is now set to "BOTH" instead of "N/A".
Trac #399, PR 62438

Keywords CHI_SQ_B and NDF_B were added to the list of keywords to be
copied from the segment B header.  Trac #394, PR 62433

New file findshift1.py:

This file contains the revised code for finding the shift of a wavecal
spectrum with respect to a template, in the dispersion direction.
Trac #394, PR 62433

Change to fpavg.py:

In createOutput, the WAVECALS and SEGMENT keywords are now deleted from the
output primary header.  Trac #395, 399, PR 62434, 62438

Changes to timetag.py:

For tagflash data, the name of the input raw file is now saved as the
value of the WAVECALS keyword in the primary header just after calling
concurrent.processConcurrentWavecal.  For auto/GO wavecals, updateFromWavecal
updates the WAVECALS keyword with the name or names of the wavecal files.
A test on wavecorr was added before calling these functions.
Trac #395, PR 62434

In initHelcorr, compute the radial velocity and assign to the V_HELIO
keyword regardless of HELCORR (but still only for spectroscopic data).
Replace function heliocentricVelocity with a version that is nearly the
same as the one in calstis.  The only difference (other than Python vs C)
is that the target coordinates will not be precessed to the time of
observation.  Trac #400, PR 62439

Revised updateFromWavecal to handle each NUV stripe separately.  Added
functions nuvPsaRegions and nuvWcaRegions for this purpose.  In doDoppcorr,
call nuvPsaRegions instead of nuvDopplerRegions, and the latter function
has been deleted.  Trac #394, PR 62433

Modified dopplerParam to ignore an FPOFFSET column that may be in the
disptab.  In the call to cosutil.getTable, use at_least_one=True instead
of exactly_one=True.  Trac #393, PR 62432

In writeCsum, changed PULSE_HEIGHT_RANGE to PHA_RANGE, because it's shorter
and probably clearer.  Added commented-out code to use pyfits.CompImageHDU
instead of ImageHDU as an example, because this change will probably be done
in the near future.  Trac #401, PR 62440

Changes to wavecal.py:

The calling sequence of storeWavecalInfo now includes the raw file name
(for segment A, if FUV); the directory will be removed, and the base file
name will be saved in the current dictionary with key "filename".
returnWavecalShift and interpolateWavecal were modified to also return a
string giving the names of the (one or two) wavecal files that were used
to get the shifts for the current observation.  Trac #395, PR 62434

In findWavecalSpectrum, the message printed to the trailer file for the
case that the spectrum was not found has been modified.  The string "----"
is printed instead of "None" as the value of the shift, and the note
"# not found in XD" is printed at the end of the line.
Trac #394, PR 62433

In findWavecalShift, check for the existence of the fp_pixel_shift column.
if it's found, set fp to 0, and add the fp_pixel_shift to the shift in the
dispersion direction for the segment or for each stripe.  Use findshift1
to find the shift in the dispersion direction.  Add x_offset to the info
dictionary.
Functions crosscor, xcStat, and ttFindWavecalShift were deleted.  crosscor
and xcStat are now in findshift1.py, and ttFindWavecalShift is no longer
needed.  Trac #393, 394, PR 62432, 62433

In findWavecalSpectrum, removed the code that checks for the existence of the
xd_range and box columns in the wcptab and assigns default values if they are
not found.  That is, these columns are now required.  Trac #332, PR 61783


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@7915 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 1090fb3d608f2e433e92f9afdf403271d9676d4d