Changes to calcos.py:
In __init__ for Observation, the case of a non-tagflash wavecal with
obstype = IMAGING was overlooked.  For this case, the data will be
calibrated as an image, without using the wavecal information.
Trac #367, PR 62091

Changes to cosutil.py:

Function setWriteToTrailer and global variable write_to_trailer were added.
This function is currently not called, but it could be used to turn on or
off writing to trailer files.  Trac #303, PR 61307


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@8137 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 8b50344d7086c55b34abf73e60bc00399fac709a