Changes to accum.py:
Function accumBasicCalibration was modified to update keyword NEXTEND.
Trac #422, PR 62939

Changes to calcos.py:

Functions openTrailerForRawInput and closeTrailerForRawInput and global
variable raw_input_trailer were added.  openTrailer and closeTrailer in
Observation now return without doing anything if raw_input_trailer is
True.  The calls to cosutil.setWriteToTrailer in readAsnTable and
dummyAsnTable have been commented out.  Trac #303, PR 61307

Function initialInfo was modified to check for a rawacq file that's not
an acq/image, for the case that the user specified a file name explicitly
(a complete file name, and no association table).  This was already done
in this function, but if the file name was given explicitly, the function
returned before reaching the test on rawacq.  Trac #422, PR 62939

In __init__ for Association, if basic_info (returned by initialInfo())
is None, just skip the current exposure, rather than not processing the
association at all.  Trac #422, PR 62939

Function setSpectrumOffset sets LAMPUSED to NONE if the lamp was not on
but LAMPUSED was not already NONE (which it shouldn't be for a wavecal).
Trac #425, PR 62997

Changes to concurrent.py:

In __init__ for NUVConcurrentWavecal, the test on opt_elem = G230L and
cenwave = 3360 has been deleted, so the segment_list will include NUVC
for all NUV data.  Trac #422, PR 62939

setShiftKeywords sets LAMPUSED to NONE if the lamp was not on but LAMPUSED
was not already NONE (which it shouldn't be).  Trac #425, PR 62997

processConcurrentWavecal was still returning (0., 0., None) for cases
where no tagflash wavecal processing was to be done.  This was left over
from when this function returned (avg_dx, avg_dy, shift1_vs_time), but
now it only returns shift1_vs_time.  Trac #422, PR 62939

Changes to extract.py:

In concatenateFUVSegments, globrt_b was added to the list of keywords
to be copied from the _b header to the _a header.  Trac #408, PR 62666

Changes to fpavg.py:

In oneInputFile, keywords segment and wavecals are deleted from the output
primary header.  This was previously done for the case that more than one
input file was averaged to the x1dsum file, but the case of just one input
file was missed.  Trac #395, 399, PR 62434, 62438

Changes to timetag.py:

In updateFromWavecal, the test on opt_elem and cenwave for G230L, 3360,
has been deleted, so the shift and dpixel keywords for strip C will not
be set to 0.  Trac #422, PR 62939


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@8201 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: ee54a603e104f76be76e04b23b7ac192cf0cb253