There was a bug in updating keyword ASN_MTYP in fltsum headers.  The keyword is only present in the SCI header, but calcos was trying to update it in the ERR and DQ headers as well.
When the shift in the dispersion direction is found for only one or two of
the NUV stripes, the shifts that are found are averaged and used to assign
values to the shifts that were not found.  Since the stripes do not move
together, a rough correction for the skew between stripes is included.
There was a sign error in this correction, however.

See Trac #386, 445 and PRs 62270, 63475.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@8274 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 729ad4fbb69a94547ea45cffb7d95dfd2a2978fa