Change to all Python modules in calcos:
Added "from __future__ import division # confidence high" to calcos,
__init__.py, setup.py, calcosparam.py, osmstep.py, and x1d.py.  Most
Python modules already had this statement, but they didn't have the
"# confidence high" comment, so that was added to all of them.
Trac #415, PR 62810

Change to accum.py:

The WAVELENGTH column is now initialized to zero in accumBasicCalibration.
Trac #481, PR 64187

Changes to concurrent.py:

In shift1Corr and shift2Corr the conditions for the upper limit (region[1])
were changed from "< region[1]" to "<= region[1]" for consistency with the
active_area array of flags in timetag.py.  There is still a discrepancy,
in that the X limits of the active area are not included, but I think it is
actually preferable to not use the X limits in this case.  Also, the integer
values 1 and 0 were changed to the boolean True and False respectively in
the calls to the 'where' function in shift1Corr and setShiftKeywords.
Trac #497, PR 64275

Change to timetag.py:

In setActiveArea, the limits of the active area were being shrunk by two
pixels on each border.  This wasn't necessary, so the active area is now
used as it is gotten from cosutil.activeArea.  An event is considered to be
within the active area if it is within the limits or exactly on a border.
Trac #497, PR 64275


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@8575 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 6cbbe7dd3cdd59375a09173ba319884b3437d790