Changes to calcos.py:
The parameter helcorr="OMIT" is now included in the call to
timetag.computeWavelengths.  Trac #481, PR 64187

Changes to concurrent.py:

In findFlash, update the lamp_is_on attribute, True if the number of flashes
is greater than zero.  This was done primarily to cover the case of "imaging
wavecals," because wavecal.ttFindWavecalSpectrum is not called for them, and
that is the function that returns the lamp_is_on flag.  Trac #425, PR 62997

Changes to phot.py:

In readImPhtTab, the values for photflam and photfnu in photdict were
replaced.  For photflam the values were given by Paul Goudfrooij in
an email message dated 2010 Jan 25.  For photfnu the values were found
by using calcphot, e.g.
calcphot "cos,nuv,mirrora,psa" "unit(4.816554456084e-18,flam)" fnu
Trac #356, PR 62062

Changes to timetag.py:

The call to doPhacorr was moved from just before the call to doRandcorr
to a point after geometric correction and the second call to setActiveArea.
Trac #513, PR 64419

A test was added to check that obstype is SPECTROSCOPIC before calling
computeWavelengths to populate the WAVELENGTH column in the corrtag table.
The heliocentric correction switch (helcorr) has been added to the calling
sequence, and if set to perform or complete this correction will be applied
to the wavelengths.  Trac #481, PR 64187


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@8689 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 955ba5b242ada92eeb156af0c1cf64e87441fc41