Relevant Trac tickets and PRs: Trac   PR ----  -----  498  64286  Calcos should compute WCS parameters for spectroscopic data.  554  65200  Calcos should allow rawtag suffix for a corrtag input file.  559  65204  Calcos should use the input exposure time in some cases.  575  65353  Calcos should have an option to find the target spectrum.
Changes to calcos.py:

Methods allScience and writeWCS in Calibration were modified to use a flag
returned by writeWCSKeywords that indicates whether the WCS keywords were
actually written.  This flag is checked before printing the name of the
SPWCSTAB and the names of the files whose headers were updated.
Trac #498, PR 64286

In Observation, the __init__ method was modified to move the check on
whether the input is a corrtag file and the input and output directories
are the same.  For corrtag input, if the file name suffix is rawtag instead
of corrtag, the input and output directories may now be the same.
Trac #554, PR 65200

The sp_nom_[abc] and sp_off_[abc] keywords were added to function
mergeKeywords.  Trac #575, PR 65353

Changes to extract.py:

xd_nominal and xd_offset were added to the calling sequence of
updateExtractionKeywords, and these values are now written to the header with
keywords SP_NOM_[abc] and SP_OFF_[abc] respectively.  In extractSegment
xd_nominal is set to the expected location of the spectrum, i.e. B_SPEC plus
SHIFT2[abc], offset to where it crosses the middle of the flt or counts image,
and xd_offset is set to the offset2 value returned by xdSearch.  In
extractCorrtag these values are currently set to -999. (i.e. not implemented
yet).  Keywords sp_nom_[abc] and sp_off_[abc] were added to the lists in
concatenateFUVSegments and copyKeywordsToInput.  Trac #575, PR 65353

Changes to spwcs.py:

Function writeWCSKeywords now returns True if the WCS keywords were actually
written and False if they weren't (they're not written if the aperture is
neither PSA nor BOA).  Trac #498, PR 64286

Changes to timetag.py:

Function recomputeExptime was modified to set the orig_exptime value in
the info dictionary to the range of times in the TIME column, rather than
what was in the EXPTIME keyword in the raw file.  This replaces the value
(from keyword EXPTIME) that was assigned in getinfo.py.
Trac #559, PR 65204

Changes to x1d.py:

The "--find" option is now included in the list of options printed by
prtOptions.  Keywords sp_nom_[abc] and sp_off_[abc] were added to the lists
in function copyKeywords.  Trac #575, PR 65353


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@9765 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 38be9c782da95f418fc408a6f97feb15ca55a77e