Changes to timetag.py:
Function mkHeaders was modified to add comment strings to the header
keywords as they are being added to the ERR and DQ extensions.
Trac #587, PR 65693

Changes to x1d.py:

Change from lists of file names to a dictionary, partly to allow easier
access to the information, but also to allow the input list to include data
from different exposures.  A printFilenames function was added, to print the
names of the input corrtag, and output flt, counts, and x1d files.
copyKeywords was modified to change keyword SP_HGT to SP_HGT_[abc].
Trac #586


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@10082 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 2a3418fdd09804661b4bcadee481d33e219515e0